Day 9: Sensor Boost

+++ src/jmh/kotlin/io/github/ephemient/aoc2019/Day2Bench.kt
@@ -15,8 +15,8 @@ open class Day2Bench {
     }
 
     @Benchmark
-    fun part1(): Int = Day2(lines).part1()
+    fun part1(): Long = Day2(lines).part1()
 
     @Benchmark
-    fun part2(): Int? = Day2(lines).part2()
+    fun part2(): Long? = Day2(lines).part2()
 }
